requirements.txt
Django==5.2.9
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
python-decouple==3.8
python-dotenv==1.2.1
psycopg2-binary==2.9.11
gunicorn==23.0.0
boto3==1.42.4
django-cors-headers==4.3.1
django-filter==24.1
